use-cases/01-simple-completion/python/pyproject.toml
[project]
name = "simple-completion"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
]
